code: rnn training
Training an rnn (using PsychRNN) with simple inputs (curvature, arc mid, arc len, fixation) to map to a behavioral output (saccade). The filenames are a bit confusing.
- nw_train_test is the main script that initiates the network parameters, etc.
- nw_task_init generates a trial for training
- nw_task_test generates a trial for testing
Note: these two are different because I want random inputs for training, but for testing I want to test the full range of curvature and arc parameters.