code and data: rnn training with cnn inputs
This is more or less the same as the one with simple inputs. The difference is that instead of generating a random curvature values between 0 and 1 for each trial, this rnn takes in a 256 dimensional vector which corresponds to the activations of pool4 of VGG16 after dimensionality reduction. Each trial now is an image and the RNN has to figure out the curvature from CNN activations. There are two sets of inputs during training. Either 1000 random shapes or 50 random shapes with 20 curvatures each. Testing both of these training types is 5 shapes with 20 curvatures each.